# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: StreamFlix API - Netflix/Amazon style streaming app
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watchlist/{itemId}:
    post:
      operationId: addToWatchlist
      tags: [watchlist]
      summary: Add item to watchlist
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "409":
          description: Already in watchlist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: removeFromWatchlist
      tags: [watchlist]
      summary: Remove item from watchlist
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /watchlist/check/{itemId}:
    get:
      operationId: checkWatchlist
      tags: [watchlist]
      summary: Check if item is in watchlist
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Watchlist status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WatchlistStatus"

  # ─── REVIEWS ─────────────────────────────────────────────────────────────────
  /reviews/{itemId}:
    get:
      operationId: getReviews
      tags: [reviews]
      summary: Get reviews for an item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create or update a review
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── ADMIN ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ── Auth ─────────────────────────────────────────────────────────────────
...
    WatchlistStatus:
      type: object
      required: [inWatchlist]
      properties:
        inWatchlist:
          type: boolean

    # ── Reviews ───────────────────────────────────────────────────────────────
    ReviewInput:
      type: object
      required: [rating]
      properties:
        rating:
          type: number
          minimum: 1
          maximum: 10
        comment:
          type: string
    Review:
      type: object
      required: [id, itemId, userId, rating, createdAt]
      properties:
        id:
          type: integer
        itemId:
          type: integer
        userId:
          type: integer
        userEmail:
          type: ["string", "null"]
        rating:
          type: number
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time

    # ── Admin ─────────────────────────────────────────────────────────────────
...